Uses cit from geocoder.rb settings instead of explicit option Refactores here lookup code to avoid overwriting whole class Refactores domain modification code Refactors here_base.rb code to DefaultLookupMethods concern Uses :here handle by default to avoid defining individual configuration for every lookup
