+++ lib/geocoder/lookups/default_lookup_methods.rb
@@ -1,14 +1,24 @@
 module Geocoder::Lookup
-  class HereBase < Base
+  module DefaultLookupMethods
+    extend ActiveSupport::Concern
+
+    def handle
+      :here
+    end
+
     def required_api_key_parts
       ["app_id", "app_code"]
     end
 
-    def name
-      fail
-    end
-
     private
+
+    def query_url_here_options(query)
+      options = {
+        app_id: api_key,
+        app_code: api_code,
+        language: (query.language || configuration.language)
+      }
+    end
 
     def api_key
       if (a = configuration.api_key)
